debug coding error
There was a minor coding error in line 84 of this code. The code was supposed to label responses as "0" if there was one or more "0"s present in the responses. Instead it labelled responses as "0" if all of the responses were "0". It is now debugged to do the former by replacing 1 by length(results_frame$y_variable[[n]]